Build with Mage and check with Revive
I like Mage for managing dependencies. It's not _so_ useful in a small
project like this, but using it here helps maintain proficiency for
larger projects. It's also useful for managing dependencies on other
binaries, like goimports and revive, so that updating the dependency
version will automatically trigger a re-install of the binary if the
local cached version doesn't match the one requested in go.mod. That's
hard to express in makefiles, but easy to check with Mage.