Use Revive rule from magehelper
I've also determined that running "go mod tidy" from within the magefile
isn't so useful, especially when it's the magefile whose code has
changed. Tidy is useful to make Go download modules that are mentioned
in go.mod but aren't present in the module cache, and we can't do that
from within the magefile if the missing modules are needed for building
the magefile in the first place.