Use nicer format for revive.toml
Revive seems to have retrofit _named_ arguments foe linters atop an
existing _positional_ argument scheme, which means it receives an array
of structs, which is awkward to express in the default syntax documented
by Revive. Turns out there's another syntax that allows for key-value
pairs and list items to appear on separate lines.