Remove type parameter on value matchers
This had always been a blemish on the interface of the matchers, but by
adding a type-erased getter method, we're now able to accept matchers of
any type.

This includes testable examples of the matchers, which ended up leading
to rewriting main implementation of the matchers as well. The previous
implementation, which was nothing but a composition of existing Gomega
matchers, turned out to generate failure messages that were not only
ugly and obtuse, but also impossible to reproduce on repeated runs,
which is a requirement for testable examples. (The messages contained
the memory addresses of the composite matchers.)

The documentation now makes it more clear what effect negation has on
the matchers. The examples demonstate the expected results.

Fixes #1.