Change underlying type to struct
I didn't like that Gomega would print the name of the type like it was a
slice (despite it being a slice after all). Now it's a struct with a
pointer. The default printing of the value seems pretty good, too.

This required (re)implementing the `BeEmpty` matcher since the one
inherited from Gomega no longer applies, but it's easy enough since it
doesn't require any type information. (The documentation claimed to
check the type, but that never happened, so the documentation has been
changed to match.)